+++ LCLandmineOutside/Data/HazardCalculationContainer.cs
@@ -1,9 +1,9 @@
 ﻿using System.Collections.Generic;
 using UnityEngine;
 
-namespace LCHazardsOutside
+namespace LCHazardsOutside.Data
 {
-    internal class HazardCalculationContainer(System.Random random, List<GameObject> spawnDenialPoints, SpawnableMapObject spawnableMapObject, int minSpawnRate, int maxSpawnRate)
+    internal class HazardCalculationContainer(System.Random random, List<GameObject> spawnDenialPoints, SpawnableMapObject spawnableMapObject, int minSpawnRate, int maxSpawnRate, int layerMask)
     {
         public System.Random Random { get; set; } = random;
         public List<GameObject> SpawnDenialPoints { get; set; } = spawnDenialPoints;
@@ -12,6 +12,7 @@ namespace LCHazardsOutside
         public int MaxSpawnRate { get; set; } = maxSpawnRate;
         public bool NeedsSafetyZone { get; set; } = false;
         public float SpawnRatioMultiplier { get; set; } = 1.5f;
+        public int LayerMask { get; set; } = layerMask;
 
-}
+    }
 }