+++ LCLandmineOutside/Data/HazardType.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace LCHazardsOutside.Data
+{
+    public enum HazardType
+    {
+        Landmine,
+        Turret,
+        SpikeRoofTrap,
+        CustomHazard
+    }
+}