+++ LCLandmineOutside/Data/MoonMinMax.cs
@@ -1,0 +1,8 @@
+﻿namespace LCHazardsOutside.Data
+{
+    public class MoonMinMax(int min, int max)
+    {
+        public int Min { get; set; } = min;
+        public int Max { get; set; } = max;
+    }
+}