+++ LCLandmineOutside/Data/SpawnStrategyType.cs
@@ -1,0 +1,9 @@
+﻿namespace LCHazardsOutside.Data
+{
+    public enum SpawnStrategyType
+    {
+        MainAndFireExit,
+        MainEntranceOnly,
+        FireExitsOnly
+    }
+}